app/internal/window: tweak minimum pixels per dp/sp
To keep text legible on low DPI monitors, the number of pixels
per dp or sp is not is not allowed to go below minDensity. However,
a minimum density of 1.25 leads to too big text and widgets on a
baseline ~96 DPI monitor. Lower the minimum to 1.0, where text is still
legible.

Signed-off-by: Elias Naur <mail@eliasnaur.com>