Revert "widget/material: propagate ButtonLayout minimum constraints to content"
This reverts commit 52ccc183b54c2a3ccc04fe69cfa08eebaab662a1.

Reason for revert:

This doesn't seem like a good idea after all. The reason for the change was to
propagate the minimum constraints to the button content. But in the simplest case,
a label, stretching the button will make the label stretch as well, leaving the label
top-aligned.

We'll revisit this issue if a real use-case comes up.